# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Persons with Login and Register",
        "description": "",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/auth/provincias": {
            "get": {
                "tags": [
                    "usuarios"
                ],
                "summary": "Listado de todes las personas",
                "operationId": "e6fd1b233f4655bff27a85896e3f3674",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "provincia": {
                                                        "type": "string",
                                                        "example": "manabí"
                                                    },
                                                    "canton": {
                                                        "type": "string",
                                                        "example": "chone"
                                                    },
                                                    "parroquia": {
                                                        "type": "string",
                                                        "example": "santarita"
                                                    },
                                                    "recinto": {
                                                        "type": "string",
                                                        "example": "uleam"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}